+++ Modules/Combat/Events/Input/ButonPressed/SlingshotButtonPressedEvent.cs
@@ -1,8 +1,5 @@
 ﻿namespace DaLion.Overhaul.Modules.Combat.Events.Input.ButonPressed;
 
-using DaLion.Overhaul;
-using DaLion.Overhaul.Modules.Combat;
-
 #region using directives
 
 using DaLion.Overhaul.Modules.Combat.Events.GameLoop.UpdateTicking;
@@ -102,7 +99,7 @@ internal sealed class SlingshotButtonPressedEvent : ButtonPressedEvent
             directionVector *= -1f;
         }
 
-        var driftVelocity = directionVector * (1f + Game1.player.addedSpeed * 0.1f) * 3f;
+        var driftVelocity = directionVector * (1f + (Game1.player.addedSpeed * 0.1f)) * 3f;
         CombatModule.State.DriftVelocity = driftVelocity;
         this.Manager.Enable<SlickMovesUpdateTickingEvent>();
     }